0029787: Avoid in presentation edges of certain continuity class
A new flag theMostAllowedEdgeClass is passed to the methods StdPrs_ShadedShape: Add, FillFaceBoundaries. It describes the most allowed continuity class of edges that will be included into the presentation. The edges with more continuity will be ignored. By default, the value is CN, i.e. all edges are included into the presentation.
The new methods in Prs3d_Drawer: SetMostContinuityClass, MostContinuityClass allows specifying the most allowed continuity class of edges for the shape's presentation. This value is taken in AIS_Shape during the compute presentation