0025748: Parallel version of progress indicator
The fix revises the progress indication mechanism. Now, the class Handle(Message_ProgressIndicator) performs the only function of calling back to user application. It just accumulates the progress provided by progress scopes.

The new class Message_ProgressScope serves to represent a scope of execution. The new instance of it is created to provide progress advancement of a new coming operation. The scopes are nested to each other to reflect the nested nature of operations.

All codes involving progress indication have been updated to new API.