0025748: Parallel version of progress indicator
The fix revises the progress indication mechanism. Now, the class Handle(Message_ProgressIndicator) performs the only function of calling back to user application. It just accumulates the progress provided by progress scopes. The counter is protected by mutex for thread-safety.

The new class Message_ProgressScope serves to represent a scope of execution. The new instance of it is created to provide progress advancement of a new coming operation. The scopes are nested to each other to reflect the nested nature of operations. Each instance of progress scope can call the method Increment of the progress indicator concurrently with other scopes, which make the work of the overall mechanism thread-safe.

All OCCT algorithms involving progress indication have been updated to new API.

Improvements in Draw_ProgressIndicator:

- Console mode has been added in order to make possible to put the progress into cout instead of draw interpreter.
- Treatment of Ctrl-Break signal has been added. Now any operation can be aborted by Ctrl-C or Ctrl-Break keystroke.