0030020: Incorrect bspline visualization
Let's we have a spline having fit points located on a direct line (let's number them 1, 2, 3), but they are located in order 1, 3, 2. There was a bug in GCPnts_TangentialDeflection that only first and last points were put in the result. Now more point is put, so that the result polygon made back movement.

The command crvtpoints has been made to accept angular value in degrees instead of radians.